Inserting Product Names and Generating Product Id
New private methods were introduced in the SalorDaoImpl Class and the
details of those methods are provided in the code itself.

I have successfully completed the insertion of product names in the
product List by creating database for each organization according to
UserId of the organization. In each database of the organization there
is a table named ProductList which contains the List of Products with
theie productIds and product Names. And according to the product Ids
table is created to manage the sales of that particular product.